refactor: 리뷰 반영 및 패키지 구조 변경 완료
