--- src/main/java/racing/ForwardConditionStrategy.java
@@ -1,5 +1,0 @@
-package racing;
-
-interface ForwardConditionStrategy {
-    boolean judgeCondition();
-}
