+++ src/main/java/racing/domain/ForwardConditionStrategy.java
@@ -1,0 +1,5 @@
+package racing.domain;
+
+public interface ForwardConditionStrategy {
+    boolean judgeCondition();
+}
