+++ src/test/java/racing/model/UnnamedCarTest.java
@@ -1,8 +1,9 @@
-package racing;
+package racing.model;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing.model.UnnamedCar;
 
 class UnnamedCarTest {
